requirements-dev.txt
-r requirements.txt
coverage
django-debug-toolbar>=1.6
pylint==2.4.4
pylint-django==2.0.14
